Rename folder name to helperMethods
Adjust the name in different files.
